pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gamewalk-helper"
version = "0.1.0"
description = "Windows walkthrough assistant using OCR/CV and voice hints."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "GameWalkthroughHelper Contributors" }]
dependencies = []

[project.optional-dependencies]
runtime = [
  "mss>=9.0.1",
  "Pillow>=10.4.0",
  "opencv-python>=4.10.0.84",
  "numpy>=2.1.0",
  "paddleocr>=2.8.1",
  "pyttsx3>=2.98",
  "PyYAML>=6.0.2",
  "requests>=2.32.0",
]

[project.scripts]
gwh = "gamewalk_helper.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

